Reverting to the state of the project at af271ebf revision 1.21.4

+++ lib/gnu_gama/g3/g3_model.cpp
@@ -242,7 +242,7 @@ void Model::update_adjustment()
 
       if (height.free())
         {
-          size_t k = U.index();
+          int k = U.index();
           height.set_index(k);
           height.add_correction(adj->x()(k));
         }
@@ -252,7 +252,7 @@ void Model::update_adjustment()
          i=par_list->begin(), e=par_list->end(); i!=e; ++i)
     {
       Parameter* p = *i;
-      if (size_t k = p->index()) p->add_correction(adj->x()(k));
+      if (int k = p->index()) p->add_correction(adj->x()(k));
     }
 
   // ..........................................................
